Forcibly update Polylang to latest version
The reason epfl-idevelop/jahia2wp has a
data/plugins/generic/polylang/polylang.zip stuck in the past is so
that things work when importing a multilingual site converted from
Jahia. Since 1) we are no longer going to do that and 2) the version
of Polylang in /wp/ has no impact whatsoever on the jahia2wp Python
script, we want to override config-lot1.yml for that one plug-in.

- Make install-plugins-and-themes.py executable

- Invoke it once more in the Dockerfile for Polylang

- Make it so "wordpress.org/plugins" is a synonym to "web" as a source
  URL for a plug-in