Work around https://github.com/ansible/ansible/issues/28674
Close the single quote of "won't" on the next line, as Ansible was
derailed by the single quote (in a comment, in a shell script that
Ansible shouldn't mess with in the first place whatsoever).

I did try to use a fancy-schmancy Unicode apostrophe, but I ran into
further Python/Unicode nonsense and gave up.