[fix] Use suitcase interpreter for the inventory script
When Ansible runs the inventory, it will use its own sys.executable
which wpsible ensures is the one from the suitcase. However, when
running the inventory directly (e.g. to look for a particular
inventory_hostname) this was previously not the case, resulting in
possible dependency problems for e.g. the `requests` module.

- Determine whether we are running the suitcase interpreter, by
comparing `realpath`s based on the well-known path starting at `__file__`
- If this is not the case, `exec` ourselves.

Fixes #277