[fix] “Prometheus StatefulSet” task failing b/c of JSON SNAFU
Historically Ansible has had two distinct features to permit
macro-expansion that results in some structured data type (other than
a string): the oldest (and most ill-advised) of the two being that it
sniffs expansion results that look like a JSON string¹, and unparses
them. There is a way to combat that behavior when it is not wanted, by
using one of the whitelisted filters as the last one in the Jinja
expression² (among which `| to_json`); unfortunately, the regex-based
detection fails to account for multi-line strings³.

In this here use case (before this commit), the use of the `>-`-style
YAML quoting doesn't, umm, actually work to eliminate embedded
newlines — In fact, `>-` is basically no different from `|-` for text
lines that are over-indented w.r.t. the first one⁴.

The TL;DR is that Ansible's `| to_json` doesn't work in case of a
multi-line, “over-”indented Jinja2 expression.

- Refactor so that the `>-` expression that matters (the one that goes
into the JSON to be presented to the k8s API) isn't multiline anymore
(i.e. no hanging indent)

- Be sure to `| to_json` in that same expression (that pipe being one
of the white-listed ones in Ansible 2.9.6 and above)

¹ https://github.com/search?q=%22def+template%22+repo%3Aansible%2Fansible+path%3Alib%2Fansible%2Ftemplate&type=Code&ref=advsearch&l=&l=

² https://docs.ansible.com/ansible/latest/reference_appendices/config.html#string-type-filters

³ https://github.com/ansible/ansible/search?q=%22self._no_type_regex%22

⁴ “folding does not apply to line breaks surrounding text lines that contain leading white space” — https://yaml.org/spec/1.2/spec.html#id2779048