[hotfix] We want `site_has_own_route` to be a Boolean
We are not entirely sure why this was working before. But if a `>` multi-line indicator is used, the final newline in the quoted multi-line string is kept; resulting in something that Ansible is well within its rights to refuse to interpret as a Boolean. As a consequence, the `when: site_has_own_route` guard in `ansible/roles/wordpress-instance/tasks/routes.yml` was always true, resulting in spurious attempts to apply the `oc` command (which doesn't - yet? - exist in the ansible-runner image).