[fix] Trigger `httpd`, `mgmt` and `cronjob` rebuild on `wp-base`'s
Not sure how this worked in the first place: did the behavior of `openshift_imagestream` allow for a more liberal kind of implicit chaining in the past? Or was that piece of the `BuildConfig` added by hand? Anyway, today we saw that the assumption that Jenkins makes (that launching a `wp-base` build, will trigger the others automatically) no longer holds.

- Set up a `triggers:` parameter to relevant `openshift_imagestream:` Ansible tasks
- Explicitly trigger on `{{ wp_base_image_name }}:latest` `imageChange`