[feature] Homemade execution environment, built on top of mgmt image
An AWX Execution Environment (EE for short) is just an ephemeral pod that AWX runs when it has “data plane”-related work to do; see explanations at https://github.com/ansible/awx/blob/devel/docs/execution_environments.md . We need to introduce our own, like we had in AWX 15.x, for the following bells and whistles:

- Run-time dependencies of our configuration-as-code: Ansible roles and collections, PHP, WordPress CLI, `new-wp-site` script etc.
- Access to NFS (/srv) under the appropriate privileges (i.e. user 33)

In this iteration, we decided to add “their” stuff on top of “our” mgmt image, instead of the other way round, again because it has become difficult to install packages on RedHat-ish images (such as, say, PHP) on account of EPERM vs. ENOSYS nonsense.

- Introduce `wp-awx-ee` BuildConfig and ImageStream
- Introduce `-t awx.build.ee` tag
- Make sure to use the per-namespace service account that has permission to run under an arbitrary UID, as decided and explained in https://jira.camptocamp.com/servicedesk/customer/portal/5/INF-5343