[unfeature] `GITHUB_API_TOKEN` in OpenShift-side builds
I decoded this token using the private key in Keybase, and found that
- it has the wrong format (does not start with `ghp_` like contemporarily-generated tokens do)
- it was expired (tested with `curl --request GET --url "https://api.github.com/octocat" --header "Authorization: Bearer $TOKEN"`)

As an aside, looking at my GitHub dashboard it is unlikely that the one in `.travis.yml` (which we cannot decode) still works either — I don't have any kind of bearer tokens (anywhere under https://github.com/settings/tokens i.e. GitHub appps, OAuth apps, fine-grained or “classic” personal access tokens) that are that old. It is tempting to conclude that the reasons our builds still work lately, is that someone has been increasing quota somewhere in the cloud.