[fix] Set `imagePullPolicy: Always` on worker containers
When one embugs, then rebuilds `wp-awx-ee`, the ImageStreamTag will update all right, but the OpenShift node where the pod will be running may already have an image with the exact same name. We need to ensure that the node will pull it again. (Thanks to Docker layer caching, impact is minimal in the no-change case.)