[kludge] Make HAProxy believe there is a pod at 128.178.222.83
As per https://access.redhat.com/solutions/6992561 it appears that
HAProxy doesn't actually use the Kubernetes `Service` mechanism directly
(as in, making TCP connects to the IP addresses of said Services).
Rather, it wants a “personal” connection to each pod, which makes
sense for finer-grained control of loadbalancing.

Trick it by creating an Endpoint. The result is that HAProxy
connect(2)s to its IPv4, as expected.