[fix] Do the RBAC on RKE2 namespaces only
Fixes https://go.epfl.ch/WPN-218 in part (the other part being
[already in `epfl-si/wp-operator`](https://github.com/epfl-si/wp-operator/commit/7fe22f63588abcb981f7a16f06901f1332b0a3a7))

On an EPFL “official” OpenShift instance with
namespace-as-a-service (that we don't have cluster-admin on), only the
[OLM controller](https://github.com/epfl-si/wp-operator/tree/main/olm)
can set up RBAC on our behalf. On the other hand, Rancher doesn't have
OLM, and thus we need to manage RBAC through Ansible there.

- Introduce `_rbac_managed` Boolean variable to discern these case
- Pepper `ansible/roles/wordpress-namespace/tasks/nginx.yml` with
  `when: _rbac_managed` constructs for all things RBAC