[feature] Multiple MariaDB
This will deploy the number of MariaDBs (servers) defined by the
variable `number_of_mariadb` in the `svc0041p-wordpress` namespace, but
only 2 in other namespaces. This is hard-coded in the `with_sequence`
Ansible parameter.

This also create the Backups objects, which are set to be ran at
the hour defined by identifier used for the naming of the MariaDB.
`mariadb-01`'s backups will be ran at 1, `mariadb-02`'s backups will be
ran at 2, etc. This is a temporary and very basic way to unsure that the
backups will not use all resources at once.