[fix] Prevent auto-updates (and more) from running on wp-cron
None of the self-modifying PHP stuff has a chance to work on Kubernetes anyway.

The initial intent of this change is to prevent what looks like a sudden
overload of the database around 22:23 CET every day.
[WP_Automatic_Updater::is_disabled()](https://github.com/WordPress/WordPress/blob/6.6-branch/wp-admin/includes/class-wp-automatic-updater.php#L33)
calls
[wp_is_file_mod_allowed](https://github.com/search?q=repo%3AWordPress%2FWordPress%20DISALLOW_FILE_MODS&type=code),
whose (default prior to applying filters) return value can be
controlled by setting `DISALLOW_FILE_MODS` to a truthy value, so there.