[feature] Kong reverse proxy plug-ins for WordPress A/B testing
The purpose of the two Kong plug-ins introduced herein, is to do A/B
testing through separate hostnames, without WordPress having to
cooperate in any way; and without needing to know which hostname(s)
we'll be using as “B” ahead of time.

Kong proxies HTTP/S requests sent to some hostname — hereinafter
called the “request hostname”; which is *not* the one that WordPress
is used to (according to its famed `siteurl` and `home` rows in the
`wp_options` SQL table). Despite that, everything should run smoothly
with no changes to either WordPress code or data; “everything” in this
case meaning rendered pages (which in turn implies that Kong must
rewrite the proxied HTML, CSS, JavaScript and JSON); HTTP
redirects (i.e. the `Location:` response header); and even Tequila
sessions, when logging in to the WordPress back-office.

We cannot do all of the above with just one plug-in, because for some
reason Kong doesn't believe in plug-ins that have both a `:response`
method (it being the only one where doing sub-requests *after* the
proxying is complete is permitted), and a `:body_filter` method (which
in turn is required to rewrite URLs in HTML).

- The `epfl-rewrite-hostnames` performs hostname substitutions both
  ways; that is, on the ingress side (`Host:` request header), as well
  as the egress side (`Location:` response header if any; and HTML body)
- The `epfl-hijack-tequila`, as the name implies, intercepts 302
  redirects to EPFL's Tequila Web SSO system, before they are sent to
  the browser. Using a Tequila client written in Lua, it creates a
  second Tequila session that will (upon successful login) result in
  the browser being redirected to the right place i.e. the request
  hostname.
- All of the above is “multi-homed”, in the sense that Kong picks up
  the request hostname from the actual HTTP request, rather than from a
  configuration file; and therefore, this proxy can be used as the
  back-end of any `Service`, `Ingress` or `Route`, regardless of which
  hostname(s) the clients will use in their HTTP traffic to Kong.