[refactor] (Soft-)eliminate `WP_SITE_NAME`
The operator is now free to use whatever nomenclature it wants for the
uploads dirname (e.g. to alias together the upload directories of two
sites, during a migration operation). However the
`/wp-data/%s/uploads` is a private implementation detail of
wp-ops (i.e. only the two serving images are coupled via it).